編集と作成機能を共通コンポーネント化
