+++ pages/work/[id].tsx
@@ -1,19 +1,28 @@
 import { Avatar, Container, Grid } from "@mui/material";
 import { GetServerSideProps } from "next";
+import { useRouter } from "next/router";
 import Breadcrumbs from "../../components/Common/Breadcrumb";
 import ChipList from "../../components/Common/ChipList";
 import MarkdownView from "../../components/Common/MarkdownView";
-import FileView from "../../components/File/FileView";
+import WorkEditor from "../../components/work/WorkEditor";
 import { WorkFileView } from "../../components/work/WorkFileView";
-import { useFile } from "../../hook/file/useFile";
 import { useWork } from "../../hook/work/useWork";
+import { WorkRequest } from "../../interfaces/work";
 
 type Props = {
   id: string;
+  modeStr?: string;
   error?: string;
 };
-const WorkDetailPage = ({ id, error }: Props) => {
+const WorkDetailPage = ({ id, modeStr, error }: Props) => {
   const { workDetail, updateWork, deleteWork } = useWork(id);
+  const router = useRouter();
+  const onSubmit = (workRequest: WorkRequest) => {
+    updateWork(workRequest).then((result) => {
+      if (!result) return;
+      router.push(`/work/${id}`);
+    });
+  };
 
   if (!workDetail) return <p>Loading...</p>;
   return (
@@ -25,30 +34,40 @@ const WorkDetailPage = ({ id, error }: Props) => {
           { text: workDetail.name },
         ]}
       />
-      <Grid>
-        <h1>{workDetail.name}</h1>
-        {workDetail.authers.map((a) => (
-          <Avatar src={a.icon_url} />
-        ))}
-        <ChipList chipList={workDetail.tags.map((t) => t.name)} />
-        <hr />
-      </Grid>
-      <Grid sx={{ marginTop: 3 }}>
-        <MarkdownView md={workDetail.description} />
-      </Grid>
-      <Grid sx={{ marginTop: 3 }}>
-        {workDetail.files.map((f) => (
-          <WorkFileView fileId={f.id} />
-        ))}
-      </Grid>
+      {modeStr === "edit" ? (
+        <>
+          <WorkEditor onSubmit={onSubmit} initWork={workDetail} />
+        </>
+      ) : (
+        <>
+          <Grid>
+            <h1>{workDetail.name}</h1>
+            {workDetail.authers.map((a) => (
+              <Avatar src={a.icon_url} />
+            ))}
+            <ChipList chipList={workDetail.tags.map((t) => t.name)} />
+            <hr />
+          </Grid>
+          <Grid sx={{ marginTop: 3 }}>
+            <MarkdownView md={workDetail.description} />
+          </Grid>
+          <Grid sx={{ marginTop: 3 }}>
+            {workDetail.files.map((f) => (
+              <WorkFileView fileId={f.id} />
+            ))}
+          </Grid>
+        </>
+      )}
     </Container>
   );
 };
 export default WorkDetailPage;
-export const getServerSideProps: GetServerSideProps = async ({ params }) => {
+export const getServerSideProps: GetServerSideProps = async ({ params, query }) => {
   try {
     const id = params?.id;
-    return { props: { id } };
+    const { mode } = query;
+    const modeStr = typeof mode === "string" ? mode : null;
+    return { props: { id, modeStr } };
   } catch (error) {
     return { props: { errors: error.message } };
   }
